medianBlur() implementation moved to separate file
